Changed recommended products from double to decimal

+++ VitalHelse/ViewModels/ReccomendationsViewModel.cs
@@ -5,8 +5,8 @@ public class ReccomendationsViewModel
     public int ProductId { get; set; }
     public string ProductName { get; set; }
     public int? StockCount { get; set; }
-    public double ProductPriceInVAT { get; set; }
-    public double? ProductCampaignPrice { get; set; }
+    public decimal ProductPriceInVAT { get; set; }
+    public decimal? ProductCampaignPrice { get; set; }
     public string? FirstImagePath { get; set; }
     public List<string> CategoryNames { get; set; } = new List<string>();
     public string ProductUrl { get; set; }
